backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Coffee",email = "hemanthmail2502@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-cloud-pubsub (>=2.29.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
